+++ router/admin.go
@@ -27,5 +27,4 @@ func init() {
 			},
 		},
 	})
-
 }
